xpi/chrome/content/test/test_model.html

<html>
<head>
	<script type="text/javascript" src="library/MochiKit.js"></script>
	<script type="text/javascript" src="library/SimpleTest.js"></script>
	<script type="text/javascript" src="test.js"></script>
	<link rel="stylesheet" type="text/css" href="library/test.css">
	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
</head>
<body>
<pre id="test">
</pre>
<script type="text/javascript; version=1.8" src="../library/00_MochiKit.js"></script>
<script type="text/javascript; version=1.8" src="../library/00_prototype.js"></script>
<script type="text/javascript; version=1.8" src="../library/00_component.js"></script>
<script type="text/javascript; version=1.8" src="../library/01_utility.js"></script>
<script type="text/javascript; version=1.8" src="../library/20_model.js"></script>
<script type="text/javascript; version=1.8" src="../library/20_Wedata.js"></script>
<script type="text/javascript; version=1.8">
try {
	// chrome://tombloo/content/test/test_model.html
	
	
	SimpleTest.waitForExplicitFinish();
	
	var d = succeed();
	
	/*
	.addBoth(function(res){
		log(res);
	})
	*/
	
	/*
	// 絶対復習
	d.addCallback(function(){
		return models['絶対復習'].post({
			item        : 'タイトル',
			body        : 'ボディ',
			description : '内容',
			itemUrl     : 'http://www.takao7.net/brushup/',
			tags        : ['memo1', 'memo2'],
		});
	})
	d.addCallback(function(){
		return models['絶対復習'].add('件名', '詳細', ['tag1', 'tag2']);
	})
	*/
	
	
	/*
	// GoogleCalendar
	d.addCallback(function(){
		return GoogleCalendar.addSimpleSchedule('開発 2009/5/10');
	});
	
	ok(GoogleCalendar.getAuthCookie());
	
	d.addCallback(function(){
		return GoogleCalendar.addSchedule('開発', '詳細', new Date('2009/5/10'));
	});
	*/
	
	
	/*
	// TextConversionServices
	d.addCallback(function(){
		return TextConversionServices.convert('ありがとう', 'ペンパン(kkshow)');
	}).addCallback(function(res){
		is(res, 'パピパポプ ');
	});
	
	d.addCallback(function(){
		return TextConversionServices.convert('ありがとう', '大阪弁');
	}).addCallback(function(res){
		is(res, 'おおきに');
	});
	*/
	
	
	/*
	// is.gd
	d.addCallback(function(){
		return models['is.gd'].shorten('http://www.example.com/');
	}).addCallback(function(url){
		is(url, 'http://is.gd/Rt');
	});
	
	d.addCallback(function(){
		return models['is.gd'].expand('http://is.gd/Rt');
	}).addCallback(function(url){
		is(url, 'http://www.example.com/');
	});
	*/
	
	
	/*
	// bit.ly
	d.addCallback(function(){
		return models['bit.ly'].shorten('http://cnn.com/');
	}).addCallback(function(url){
		is(url, 'http://bit.ly/eBZKT');
	});
	d.addCallback(function(){
		return models['bit.ly'].expand('http://bit.ly/eBZKT');
	}).addCallback(function(url){
		is(url, 'http://cnn.com/');
	});
	
	d.addCallback(function(){
		return models['j.mp'].shorten('http://cnn.com/');
	}).addCallback(function(url){
		is(url, 'http://j.mp/eBZKT', 'j.mp: shorten');
	});
	d.addCallback(function(){
		return models['j.mp'].expand('http://j.mp/eBZKT');
	}).addCallback(function(url){
		is(url, 'http://cnn.com/', 'j.mp: expand');
	});
	*/
	
	
	/*
	// FriendFeed
	d.addCallback(function(){
		return FriendFeed.post({
			type        : 'photo',
			pageUrl     : 'http://www.goo.ne.jp/',
			page        : 'goo',
			itemUrl     : 'http://www.goo.ne.jp/img2/logo/gh_logo2.gif',
			description : 'ポータルサイト',
		});
	});
	*/
	
	
	/*
	// Instapaper
	d.addCallback(function(){
		return Instapaper.post({
			item        : 'goo',
			itemUrl     : 'http://dictionary.goo.ne.jp/',
			description : '辞書',
		});
	});
	*/
	
	
	/*
	// YahooBookmarks
	YahooBookmarks.post({
		pageUrl     : 'http://www.goo.ne.jp/',
		itemUrl     : 'http://dictionary.goo.ne.jp/',
		item        : 'goo',
		description : '今日のニュース',
		tags        : ['search', 'portal'],
		private     : true,
	});
	
	d.addCallback(function(){
		return YahooBookmarks.getSuggestions('http://www.hatena.ne.jp/');
	});
	d.addCallback(function(res){
		log(res);
		
		ok(res.tags.length);
		ok(res.popular.length);
		
		var tag = res.tags[0];
		ok(tag.name.length > 0, 'YahooBookmarks: タグの名前を取得できているか');
	});
	*/
	
	/*
	// Yahoo
	d.addCallback(function(){
		return Yahoo.getKanaReadings('"写真","あとで読む"');
	}).addCallback(function(res){
		is(res.join(''), '"しゃしん","あとでよむ"');
	});
	d.addCallback(function(){
		return Yahoo.getRomaReadings('"写真","あとで読む"');
	}).addCallback(function(res){
		is(res.join(''), '"shashinn","atodeyomu"');
	});
	*/
	
	
	/*
	// Kawa
	d.addCallback(function(){
		return Kawa.getRomaReadings('"写真","あとで読む"');
	}).addCallback(function(res){
		log(res);
	});
	*/
	
	// Hatena
	/*
	d.addCallback(function(){
		return Hatena.login('user', 'password');
	});
	d.addCallback(function(){
		return Hatena.logout();
	});
	
	d.addCallback(function(){
		return Hatena.getCurrentUser();
	}).addCallback(function(user){
		ok(user, 'Hatena.getCurrentUser');
	});
	
	d.addCallback(function(){
		return Hatena.getToken();
	}).addCallback(function(token){
		ok(token, 'Hatena.getToken');
	});
	*/
	
	/*
	// HatenaBookmark
	ok(Hatena.getPasswords().length, 'Hatena: アカウント情報を取得できているか');
	ok(Hatena.getAuthCookie(), 'Hatena: ログインしているか(未ログインの場合エラー)');
	
	d.addCallback(function(){
		return HatenaBookmark.getToken();
	}).addCallback(function(token){
		log(token);
		ok(token);
	});
	
	d.addCallback(function(){
		return HatenaBookmark.getSuggestions('http://jp.msn.com/');
	}).addCallback(function(res){
		ok(res.tags.length, 'HatenaBookmark.getSuggestions(tags)');
		ok(res.recommended.length, 'HatenaBookmark.getSuggestions(recommended)');
	});
	
	d.addCallback(bind('post', HatenaBookmark), {
		itemUrl     : 'http://dictionary.goo.ne.jp/',
		description : '辞書',
		tags        : ['search', 'portal'],
	});
	d.addCallback(function(){
		ok(true, 'HatenaBookmark: 追加');
	});
	*/
	
	
	/*
	// HatenaStar
	d.addCallback(function(){
		return HatenaStar.post({
			pageUrl     : 'http://www.goo.ne.jp/',
			itemUrl     : 'http://dictionary.goo.ne.jp/',
			item        : 'goo',
			description : '今日のニュース',
		});
	})
	d.addCallback(function(){
		return HatenaStar.remove({
			itemUrl     : 'http://dictionary.goo.ne.jp/',
			description : '今日のニュース',
		});
	})
	*/
	
	
	/*
	// GoogleBookmarks
	d.addCallback(function(){
		return GoogleBookmarks.post({
			itemUrl     : 'http://www.goo.ne.jp/',
			item        : 'goo',
			description : 'ポータルサイト',
			tags        : ['search', 'portal'],
		});
	})
	*/
	
	
	/*
	// Twitter
	d.addCallback(function(){
		return Twitter.getToken();
	});
	d.addCallback(function(token){
		ok(token.authenticity_token > 3);
	});
	
	d.addCallback(function(){
		return Twitter.addFavorite(808614559);
	});
	d.addCallback(function(){
		return Twitter.post({
			itemUrl     : 'http://www.goo.ne.jp/',
			item        : 'goo',
			description : 'ポータルサイト',
		});
	});
	
	d.addCallback(function(){
		return Twitter.remove(847473592);
	});
	d.addCallback(function(){
		return Twitter.getRecipients();
	});
	d.addCallback(function(recipients){
		ok(recipients.length > 3);
		ok(recipients[0].name);
	});
	*/
	
	
	/*
	//Plurk
	d.addCallback(function(){
		return Plurk.post({
			description : 'こんにちは',
		});
	});
	d.addCallback(function(res){
		log(res);
	});
	*/
	
	
	/*
	// Pinboard
	d.addCallback(function(){
		return Pinboard.post({
			itemUrl     : 'http://www.goo.ne.jp/',
			item        : 'goo',
			description : 'ポータルサイト',
			tags        : ['search', 'portal', 'test'],
		});
	});
	
	d.addCallback(function(){
		return Pinboard.getUserTags();
	});
	d.addCallback(function(tags){
		var tag = tags[0];
		ok(tag.name.length > 0, 'Pinboard: タグの名前を取得できているか');
	});
	
	d.addCallback(function(){
		return Pinboard.getRecommendedTags('http://tumblr.com/');
	});
	d.addCallback(function(tags){
		ok(tags, 'Pinboard: お勧めタグが取得できるか');
	});
	
	d.addCallback(function(){
		return Pinboard.getSuggestions('http://msn.com/');
	});
	d.addCallback(function(suggestions){
		ok(suggestions.form.tags);
		ok(suggestions.recommended);
	});
	*/
	
	
	/*
	// Delicious
	d.addCallback(function(){
		return Delicious.post({
			itemUrl     : 'http://www.goo.ne.jp/',
			item        : 'goo',
			description : 'ポータルサイト',
			tags        : ['search', 'portal'],
			private     : true,
		});
	});
	
	d.addCallback(bind('getUserTags', Delicious));
	d.addCallback(function(tags){
		// 2011/9/28 Delicious移行に伴って動作せず(フィード再開待ち)
		var tag = tags[0];
		ok(tag.name.length > 0, 'Delicious: タグの名前を取得できているか');
		ok(tag.frequency!=null, 'Delicious: タグの利用回数を取得できているか');
	});
	
	d.addCallback(function(){
		return Delicious.getSuggestions('http://techcrunch.com/');
	});
	d.addCallback(function(suggestions){
		ok(suggestions.recommended.length, 'Delicious: お勧めタグなどが取得できるか');
	});
	
	d.addCallback(function(){
		return Delicious.getCurrentUser();
	});
	d.addCallback(function(user){
		ok(user, 'Delicious: ログイン中のユーザー名を取得できるか(JSON取得)');
	});
	*/
	
	
	/*
	// GoogleWebHistory
	d.addCallback(bind('post', GoogleWebHistory), 'http://google.com/');
	d.addCallback(function(){
		ok(true, 'GoogleWebHistory: 追加');
	});
	*/
	is(GoogleWebHistory.getCh('http://google.com'), '6-1226505498');
	
	
	/*
	// Flickr
	d.addCallback(bind('addFavorite', Flickr), 2371304094);
	d.addCallback(function(){
		ok(true, 'Flickr Favorite 追加');
	});
	
	d.addCallback(bind('removeFavorite', Flickr), 2371304094);
	d.addCallback(function(){
		ok(true, 'Flickr Favorite 削除');
	});
	
	d.addCallback(bind('getInfo', Flickr), 2052869076);
	d.addCallback(function(info){
		is(info.title._content, 'Mya');
		is(info.owner.username, 'Ben Pearce');
		is(info.id,             '2052869076');
		is(info.owner.nsid,     '14227047@N05');
		is(info.originalsecret, 'a02938d5af');
		is(info.secret,         'd7852e1a0b');
		is(info.server,         '2106');
		
		is(info.urls.url[0].type,     'photopage');
		is(info.urls.url[0]._content, 'http://www.flickr.com/photos/benpearce/2052869076/');
	});
	d.addCallback(bind('getSizes', Flickr), 2052869076);
	d.addCallback(function(sizes){
		is(sizes.length, 5);
		is(sizes[0].height, 75);
		is(sizes[0].width, 75);
		is(sizes[0].label,  'Square');
		is(sizes[0].source, 'http://farm3.static.flickr.com/2106/2052869076_d7852e1a0b_s.jpg');
		is(sizes[0].url,    'http://www.flickr.com/photos/benpearce/2052869076/sizes/sq/');
	});
	d.addCallback(bind('getSizes', Flickr), 2150907410);
	d.addErrback(function(e){
		is(e.message, 'Photo not found');
	});
	
	d.addCallback(function(){
		return Flickr.getToken();
	}).addCallback(function(token){
		ok(token);
	});
	
	ok(Flickr.getAuthCookie(), 'Flickr: ログインしているか(未ログインの場合エラー)');
	
	d.addCallback(function(){
		return Flickr.post({
			file    : getLocalFile('chrome://tombloo/skin/quote.png'),
			page    : '',
			private : true,
			tags    : ['test', 'icon'],
		});
	}).addCallback(function(id){
		ok(id, 'Flickr: メタ情報を付加してアップロード');
	});
	
	d.addCallback(function(){
		return Flickr.upload({
			file : getLocalFile(Flickr.ICON),
		});
	}).addErrback(function(err){
		is(err.message, 'Filetype was not recognised', 'Flickr: アップロードエラーメッセージの確認');
	});
	*/
	
	
	/*
	// FFFFOUND
	d.addCallback(bind('getToken', FFFFOUND));
	d.addCallback(function(token){
		ok(token);
	});
	
	d.addCallback(bind('post', FFFFOUND), {
		item    : 'はてな',
		itemUrl : 'http://www.hatena.ne.jp/images/top/navi_icon.gif',
		pageUrl : 'http://www.hatena.ne.jp/',
	}).addCallbacks(partial(ok, false), function(e){
		log(e);
		is(e.message, 'The image you choose is too small.', 'ポスト失敗: 画像小さい')
	});
	d.addCallback(bind('post', FFFFOUND), {
		item    : 'はてな',
		itemUrl : 'http://www.hatena.ne.jp/',
		pageUrl : 'http://www.hatena.ne.jp/',
	}).addCallbacks(partial(ok, false), function(e){
		is(e.message, 'UNSUPPORTED_TYPE', 'ポスト失敗: サポート外のタイプ')
	});
	
	d.addCallback(bind('post', FFFFOUND), {
		item    : 'Mya',
		itemUrl : 'http://farm3.static.flickr.com/2106/2052869076_a02938d5af_o.jpg',
		pageUrl : 'http://www.flickr.com/photos/benpearce/2052869076/sizes/o/',
	});
	
	d.addCallback(bind('remove', FFFFOUND), 159182);
	
	d.addCallback(bind('iLoveThis', FFFFOUND), 159322);
	d.addCallback(bind('iLoveThis', FFFFOUND), 159322).addCallbacks(
		partial(ok, false), 
		function(e){
			is(e.message, 'EXISTS', 'ポスト失敗: 重複')
		});
	*/
	
	
	/*
	// Wassr
	d.addCallback(function(){
		return Wassr.post({
			itemUrl     : 'http://www.goo.ne.jp/',
			item        : 'goo',
			description : 'ポータルサイト',
		});
	});
	*/
	
	
	/*
	// Jaiku
	d.addCallback(function(){
		return Jaiku.post({
			itemUrl     : 'http://www.goo.ne.jp/',
			item        : 'goo',
			description : 'ポータルサイト',
		});
	});
	
	ok(Jaiku.getCurrentUser().length > 1, 'Jaiku: ユーザー名を取得できるか(未ログインの場合エラー)');
	*/
	
	is(Gyazo.getId().length, 14);
	
	/*
	Mento
	d.addCallback(function(){
		return Mento.save({
			title       : 'goo 辞書', 
			url         : 'http://dictionary.goo.ne.jp/', 
			image       : 'http://www.goo.ne.jp/img2/logo/gh_logo2.gif',
			tags        : ['便利', 'dictionary', 'english'], 
			description : '辞書',
			private     : false,
		});
	});
	*/
	
	
	/*
	//LibraryThing
	d.addCallback(function(){
		return LibraryThing.post({
			itemUrl : 'http://amazon.co.jp/o/ASIN/4334032060',
		});
	});
	*/
	
	
	/*
	d.addCallback(function(){
		return models['Sharebee.com'].decrypt('0028a325');
	}).addCallback(function(res){
		ok(res.links.length);
		ok(res.fileName);
	});
	*/
	
	
	/*
	d.addCallback(function(){
		return LivedoorClip.getToken();
	});
	d.addCallback(function(token){
		ok(token);
	});
	d.addCallback(function(){
		return LivedoorClip.post({
			rate        : 5,
			private     : true,
			item        : 'goo 辞書',
			itemUrl     : 'http://dictionary.goo.ne.jp/',
			description : '辞書',
			body        : '大辞林 第二版',
			tags        : ['便利', 'dictionary', 'english'],
		});
	});
	
	d.addCallback(function(){
		return LivedoorClip.getSuggestions('http://www.hatena.ne.jp/');
	});
	d.addCallback(function(res){
		ok(res.tags.length);
	});
	ok(LivedoorClip.getAuthCookie(), 'LivedoorClip: ログインしているか(未ログインの場合エラー)');
	*/
	
	/*
	// 8tracks
	d.addCallback(function(){
		return models['8tracks'].getPlayToken();
	});
	d.addCallback(function(token){
		ok(token);
	});
	
	d.addCallback(function(){
		return models['8tracks'].getPlaylist(26179);
	});
	d.addCallback(function(list){
		is(list.length, 15);
		
		var track = list[0];
		is(track.mixName, '!!!may.tyo!!!');
		is(track.mixDescription, 'tokyo');
		is(track.photo, 'http://cf3.8tracks.us/mix_covers/000/026/179/26235.sq56.jpg');
		is(track.referenceUrl, 'http://8tracks.com/isbsh/maytyo');
		
		ok(/^http.*\.mp3/.test(track.item));
		ok(track.title);
		ok(track.trackId);
		ok(track.contributor);
		is(track.allowSkip, 'true');
	});
	*/
	
	
	// FirefoxBookmark
	var uri = 'chrome://tombloo/content/prefs.xul';
	var title = 'Tombloo 設定画面';
	var description = 'DESCRIPTION';
	var params = {
		title : title,
		uri : uri,
		description : description,
	};
	
	FirefoxBookmark.addBookmark(uri, title, ['tombloo', 'chrome'], description);
	sameObject(FirefoxBookmark.getBookmark(uri), params, 'FirefoxBookmark: ブックマークの追加');
	ok(FirefoxBookmark.isBookmarked(uri));
	
	FirefoxBookmark.removeBookmark(uri);
	ok(!FirefoxBookmark.getBookmark(uri), 'FirefoxBookmark: ブックマークの削除');
	ok(!FirefoxBookmark.isBookmarked(uri));
	
	FirefoxBookmark.addBookmark(params);
	ok(FirefoxBookmark.isBookmarked(uri), 'FirefoxBookmark: ブックマークの追加(ハッシュ形式)');
	FirefoxBookmark.removeBookmark(uri);
	
	FirefoxBookmark.createFolder('TEST', NavBookmarksService.bookmarksMenuFolder);
	
	var folderId = FirefoxBookmark.getFolder('TEST', NavBookmarksService.bookmarksMenuFolder);
	ok(folderId, 'FirefoxBookmark: フォルダの作成');
	FirefoxBookmark.removeItem(folderId);
	folderId = FirefoxBookmark.getFolder('TEST');
	ok(! folderId, 'FirefoxBookmark: フォルダの削除/フォルダIDの省略');
	
	
	// NDrive
	is(NDrive.incrementFile('/test(2).jpg'), '/test(3).jpg');
	is(NDrive.incrementFile('test(100).jpg'), 'test(101).jpg');
	is(NDrive.incrementFile('te.st(100).jpg'), 'te.st(101).jpg');
	is(NDrive.incrementFile('test(2)'), 'test(3)');
	is(NDrive.incrementFile('/a/b/c/test(2)'), '/a/b/c/test(3)');
	
	is(NDrive.incrementFile('test'), 'test(2)');
	is(NDrive.incrementFile('test.jpg'), 'test(2).jpg');
	is(NDrive.incrementFile('/a/b/c/test'), '/a/b/c/test(2)');
	is(NDrive.incrementFile('/a/b/c/test.jpg'), '/a/b/c/test(2).jpg');
	is(NDrive.incrementFile('/a.1/b.2/c.3/te.st.jpg'), '/a.1/b.2/c.3/te.st(2).jpg');
	
	is(NDrive.validateFileName('<t|e:s/t>.jpg'), '(t-e-s-t).jpg');
	is(NDrive.validateFileName('"t\\e?s*t".jpg'), "'t-e-s-t'.jpg");
	
	
	// Soundcloud
	var url = 'http://soundcloud.com/user/track';
	is(Soundcloud.normalizeTrackUrl(url), url);
	is(Soundcloud.normalizeTrackUrl(url + '/'), url);
	is(Soundcloud.normalizeTrackUrl(url + '/download'), url);
	
	
	d.addErrback(function(e){
		error(e.message);
		ok(false, '遅延テストエラー');
	});
	d.addBoth(function(){
		SimpleTest.finish()
	});
	
} catch (err) {
	var s = [];
	s.push('TEST SUITE FAILURE!');
	for(var prop in err)
		s.push(prop + ':  ' + err[prop]);
	ok(false, s.join('\n'));
	
	SimpleTest.finish()
}
</script>

</body>
</html>
